[wallet-ext] Rework recent tx loading (#7846)
