+++ apps/wallet/src/ui/app/redux/RootReducer.ts
@@ -9,14 +9,12 @@ import permissions from './slices/permissions';
 import suiObjects from './slices/sui-objects';
 import transactionRequests from './slices/transaction-requests';
 import transactions from './slices/transactions';
-import txresults from './slices/txresults';
 
 const rootReducer = combineReducers({
     account,
     app,
     suiObjects,
     transactions,
-    txresults,
     permissions,
     transactionRequests,
 });
